backend/requirements.txt
# Core Django
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1

# Database
psycopg2-binary==2.9.11

# API Documentation
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1

# CORS and Security
django-cors-headers==4.9.0

# Caching and Task Queue
django-redis==6.0.0
redis==6.4.0
celery==5.4.0
django-celery-beat==2.7.0

# Google OAuth
google-auth==2.41.1
google-auth-oauthlib==1.2.2

# Image handling
pillow==11.3.0

# Configuration
python-decouple==3.8

# HTTP requests
requests==2.32.5

# Utilities
PyYAML==6.0.3
